Finish logic and basic GUI
Finished the coding necessary and fixed an issue where justify-items
was used instead of justify-content in a flexbox.

Also fixed an issue where clicking on a button would result in the
appropriate data-selection value, however, if the user clicked on the
image inside the button it would result in NaN.

Rearranged variables and added a simple init function to keep things
neat and tidy.